Begin work on: https://github.com/elasticsearch/elasticsearch/issues/5836
Implements a ```"lenient"``` property for each ```{ "highlight" }``` field, where if
set to ```true``` the highlighting process will silently skip fields that aren't
mapped with ```index_options``` compatible with the specified highlighter ```type```.

For example, you can now do this:
```json

"highlight": {
    "pre_tags": ["<em>"],
    "post_tags": ["</em>"],
    "fields": {
        "*": {
            "highlight_query": {
                "match_phrase": {
                    "_all": "1977-03-20"
                }
            },
            "number_of_fragments": 0,
            "type": "fvh",
            "lenient": true
        }
    },
    "require_field_match": false
}
```

To highlight any field that's compatible with the ```fvh``` highlighter that has the phrase 1997-03-20.
(note the ```"*"``` for the field name)

If a field isn't indexed in a compatible way, it simply isn't highlighted.

The ```lenient``` property can also be set globally in the ```"highlight"``` section so that it applies
to all listed fields, and the default, if not specified at all, is ```false``` to maintain backwards-
compatibility.

Unit tests don't exist yet because when I opened up HighlighterSearchTests.java my head asploded!